refactor(linter/plugins): move `definePlugin` and `defineRule` into own file (#16169)
Pure refactor. Move `definePlugin` and `defineRule` into their own module. Rule tester will live in this directory too.
